Allow "on" as truthy value in YAML linting (#4)
This change modifies the `.yamllint.yaml` configuration file to allow "on" as a valid truthy value in addition to "true" and "false". The `truthy` rule is updated to include these allowed values while maintaining its warning level. This adjustment provides more flexibility in YAML files that may use "on" as a valid boolean representation.